# Performance: fix hydration issues on find wallets page