--- a/src/components/FindWallet/WalletTable/SupportedLanguagesTooltip.tsx
+++ b/src/components/FindWallet/WalletTable/SupportedLanguagesTooltip.tsx
@@ -1,22 +1,27 @@
-import { Box, Text } from "@chakra-ui/react"
+import { Text } from "@chakra-ui/react"
 
 import Tooltip from "@/components/Tooltip"
 
-import { formatSupportedLanguages } from "@/lib/utils/wallets"
+import { formatStringList } from "@/lib/utils/wallets"
 
 import { NUMBER_OF_SUPPORTED_LANGUAGES_SHOWN } from "@/lib/constants"
 
 type SupportedLanguagesTooltipProps = {
   supportedLanguages: string[]
-  restText: string
 }
 
 // Tooltip to show other supported languages on a wallet
 export const SupportedLanguagesTooltip = ({
   supportedLanguages,
-  restText,
 }: SupportedLanguagesTooltipProps) => {
-  const tooltipContent = formatSupportedLanguages(
+  const numberOfSupportedLanguages = supportedLanguages.length
+  const rest = numberOfSupportedLanguages - NUMBER_OF_SUPPORTED_LANGUAGES_SHOWN
+
+  if (rest <= 0) {
+    return null
+  }
+
+  const tooltipContent = formatStringList(
     supportedLanguages.slice(NUMBER_OF_SUPPORTED_LANGUAGES_SHOWN)
   )
 
@@ -28,7 +33,7 @@ export const SupportedLanguagesTooltip = ({
         fontSize="md !important"
         fontWeight="normal !important"
       >
-        {restText}
+        + {rest}
       </Text>
     </Tooltip>
   )