--- a/src/components/FindWallet/WalletTable/index.tsx
+++ b/src/components/FindWallet/WalletTable/index.tsx
@@ -1,5 +1,4 @@
 import { ReactNode, useContext } from "react"
-import { useRouter } from "next/router"
 import { useTranslation } from "next-i18next"
 import { MdExpandLess, MdExpandMore } from "react-icons/md"
 import {
@@ -17,7 +16,7 @@ import {
   Text,
 } from "@chakra-ui/react"
 
-import { ChildOnlyProp, WalletData, WalletFilter } from "@/lib/types"
+import type { ChildOnlyProp, Wallet, WalletFilter } from "@/lib/types"
 
 import { ButtonLink } from "@/components/Buttons"
 import { WalletMoreInfo } from "@/components/FindWallet/WalletTable/WalletMoreInfo"
@@ -27,8 +26,7 @@ import Tag from "@/components/Tag"
 
 import { trackCustomEvent } from "@/lib/utils/matomo"
 import {
-  formatSupportedLanguages,
-  getSupportedLanguages,
+  formatStringList,
   getWalletPersonas,
   walletsListingCount,
 } from "@/lib/utils/wallets"
@@ -118,7 +116,7 @@ const WalletContentHeader = (props: ChildOnlyProp) => (
   />
 )
 
-const Wallet = forwardRef<ChildOnlyProp, "tr">((props, ref) => (
+const WalletRow = forwardRef<ChildOnlyProp, "tr">((props, ref) => (
   <Grid
     tabIndex={0}
     ref={ref}
@@ -202,7 +200,7 @@ export type WalletTableProps = {
   filters: WalletFilter
   resetFilters: () => void
   resetWalletFilter: React.MutableRefObject<() => void>
-  walletData: WalletData[]
+  walletData: Wallet[]
   onOpen: () => void
 }
 
@@ -214,7 +212,6 @@ const WalletTable = ({
   onOpen,
 }: WalletTableProps) => {
   const { t } = useTranslation("page-wallets-find-wallet")
-  const { locale } = useRouter()
   const {
     featureDropdownItems,
     filteredWallets,
@@ -299,21 +296,6 @@ const WalletTable = ({
           const hasDeviceLabels = deviceLabels.length > 0
           const hasAllLabels = hasPersonasLabels && hasDeviceLabels
 
-          // Supported languages
-          const supportedLanguages = getSupportedLanguages(
-            wallet.languages_supported,
-            locale!
-          )
-          const numberOfSupportedLanguages = supportedLanguages.length
-          const sliceSize = NUMBER_OF_SUPPORTED_LANGUAGES_SHOWN
-          const rest = numberOfSupportedLanguages - sliceSize
-          const restText = `${rest > 0 ? "+" : ""} ${rest > 0 ? rest : ""}`
-
-          const formattedSupportedLanguages = formatSupportedLanguages(
-            supportedLanguages,
-            sliceSize
-          )
-
           const showMoreInfo = (wallet) => {
             updateMoreInfo(wallet.key)
             // Log "more info" event only on expanding
@@ -330,7 +312,7 @@ const WalletTable = ({
               key={wallet.key}
               bg={wallet.moreInfo ? "background.highlight" : "transparent"}
             >
-              <Wallet
+              <WalletRow
                 // Make wallets more info section open on 'Enter'
                 onKeyUp={(e) => {
                   if (e.key === "Enter") showMoreInfo(wallet)
@@ -414,13 +396,13 @@ const WalletTable = ({
                                 fontWeight="normal !important"
                               >
                                 {/* Show up to 5 supported languages and use a tooltip for the rest */}
-                                {`${formattedSupportedLanguages}`}{" "}
-                                {rest > 0 && (
-                                  <SupportedLanguagesTooltip
-                                    supportedLanguages={supportedLanguages}
-                                    restText={restText}
-                                  />
-                                )}
+                                {`${formatStringList(
+                                  wallet.supportedLanguages,
+                                  NUMBER_OF_SUPPORTED_LANGUAGES_SHOWN
+                                )}`}{" "}
+                                <SupportedLanguagesTooltip
+                                  supportedLanguages={wallet.supportedLanguages}
+                                />
                               </Text>
                             </Flex>
                           </Stack>
@@ -474,7 +456,7 @@ const WalletTable = ({
                     </ButtonLink>
                   </Box>
                 </Box>
-              </Wallet>
+              </WalletRow>
 
               {wallet.moreInfo && (
                 <WalletMoreInfo