--- a/src/lib/types.ts
+++ b/src/lib/types.ts
@@ -581,6 +581,10 @@ export interface WalletData {
   new_to_crypto?: boolean
 }
 
+export interface Wallet extends WalletData {
+  supportedLanguages: string[]
+}
+
 export type WalletFilter = typeof WALLETS_FILTERS_DEFAULT
 
 export interface WalletFilterData {