--- a/src/lib/utils/wallets.ts
+++ b/src/lib/utils/wallets.ts
@@ -12,7 +12,7 @@ import {
   NEW_TO_CRYPTO_FEATURES,
   NFTS_FEATURES,
 } from "../constants"
-import { WalletData, WalletFilter } from "../types"
+import type { WalletData, WalletFilter } from "../types"
 
 export const getSupportedLocaleWallets = (locale: string) =>
   shuffle(
@@ -65,21 +65,21 @@ export const getWalletPersonas = (wallet: WalletData) => {
 
 // Get a list of wallet supported languages with native title
 export const getSupportedLanguages = (
-  walletSupportedLanguages: string[],
+  supportedLanguageCodes: string[],
   locale: string
 ) => {
   const supportedLanguages = [] as string[]
 
   // current locale should appear first on the list, this manipulates the array to move it to the top if needed
   const supportsCurrentLocale = (current) => current === locale
-  const localeIndex = walletSupportedLanguages.findIndex(supportsCurrentLocale)
+  const localeIndex = supportedLanguageCodes.findIndex(supportsCurrentLocale)
 
   if (localeIndex >= 0) {
-    walletSupportedLanguages.splice(localeIndex, 1)
-    walletSupportedLanguages.unshift(locale)
+    supportedLanguageCodes.splice(localeIndex, 1)
+    supportedLanguageCodes.unshift(locale)
   }
 
-  walletSupportedLanguages.forEach((supportedLanguage) => {
+  supportedLanguageCodes.forEach((supportedLanguage) => {
     // Get supported language name
     const supportedLanguageName = getLanguageCodeName(supportedLanguage, locale)
     // Capitalize supported language name
@@ -90,13 +90,8 @@ export const getSupportedLanguages = (
 }
 
 // Format languages list to be displayed on UI label
-export const formatSupportedLanguages = (
-  supportedLanguages: string[],
-  sliceSize?: number
-) => {
-  return sliceSize
-    ? supportedLanguages.slice(0, sliceSize).join(", ")
-    : supportedLanguages.join(", ")
+export const formatStringList = (strings: string[], sliceSize?: number) => {
+  return sliceSize ? strings.slice(0, sliceSize).join(", ") : strings.join(", ")
 }
 
 // Get border custom color for Persona filter