<!--- Provide a general summary of your changes in the Title above -->

## Description

Fixes hydration issues by:
- sorting the wallets list at build time
- replacing the table/th/tr/td components with divs because it was throwing invalid html errors
- format the languages support for each wallet at build time

## Tests results
Main improvement in Task 2 (first render), nearly 1 second faster using CPU 4x slowdown + slow 4g connection.

Current dev
- Task 1: 500ms
- **Task 2: 3s**
- Task 3: 1.5s

Fixing hydration issue
- Task 1: 500ms
- **Task 2: 2s**
- Task 3: 1.2s

![image](https://github.com/ethereum/ethereum-org-website/assets/468158/7d580e6e-d1cd-414d-bac8-66b0fd1947d1)

## Preview URL
- https://deploy-preview-13066--ethereumorg.netlify.app/en/wallets/find-wallet