@pettinarip 2024-06-11
> @pettinarip Am I correct that this would update with a new sort order every time the site is built/deployed?
> 
> This would put it around a 2-week cadence; have we considered (or tried) the `revalidate` flag to tell the server to fetch this again periodically (maybe once a day)? Don't see that as a blocker, but would be nice to increase the frequency imo.

Great point. Completely forgot to add that....https://github.com/ethereum/ethereum-org-website/pull/13138